changed directory of update APIs to user
