+++ server/models/user/index.js
@@ -1,0 +1,11 @@
+const getUser = require('./getUser')
+const updateUser = require('./updateUser')
+const updatePassword = require('./updatePassword')
+const updateProfileImage = require('./updateProfileImage')
+
+module.exports = {
+  getUser,
+  updateUser,
+  updatePassword,
+  updateProfileImage,
+}
