+++ server/models/utils/index.js
@@ -1,3 +1,7 @@
 const sendEmail = require('./email')
+const isPasswordCorrect = require('./isPasswordCorrect')
 
-module.exports = { sendEmail }
+module.exports = {
+  sendEmail,
+  isPasswordCorrect,
+}
