+++ server/routes/auth/index.js
@@ -3,11 +3,7 @@ const verifyEmailRouter = require('./verifyEmail')
 const loginRouter = require('./login')
 const forgotPasswordRouter = require('./forgotPassword')
 const resetPassowrdRouter = require('./resetPassword')
-const updatePasswordRouter = require('./updatePassword')
-const updateUserRouter = require('./updateUser')
 const verifyNewEmailRouter = require('./verifyNewEmail')
-const getUserRouter = require('./getUser')
-const uploadProfileImage = require('./uploadProfileImage')
 
 module.exports = {
   signupRouter,
@@ -15,9 +11,5 @@ module.exports = {
   loginRouter,
   forgotPasswordRouter,
   resetPassowrdRouter,
-  updatePasswordRouter,
-  updateUserRouter,
   verifyNewEmailRouter,
-  getUserRouter,
-  uploadProfileImage,
 }
