+++ server/routes/user/index.js
@@ -1,0 +1,11 @@
+const getUserRouter = require('./getUser')
+const updateUserRouter = require('./updateUser')
+const updatePasswordRouter = require('./updatePassword')
+const updateProfileImageRouter = require('./updateProfileImage')
+
+module.exports = {
+  getUserRouter,
+  updateUserRouter,
+  updatePasswordRouter,
+  updateProfileImageRouter,
+}
